# 3x Faster Text Preprocessing